+++ src/transactions.rs
@@ -9,30 +9,38 @@ use crate::utils::blake2_256;
 use crate::{rlp_encodable, U256};
 use secp256k1::ecdsa::{RecoverableSignature, RecoveryId};
 use secp256k1::{Message, PublicKey, Secp256k1};
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
 
 rlp_encodable! {
     /// Represents a single VeChain transaction.
+    #[cfg_attr(feature="serde", serde_with::serde_as)]
+    #[cfg_attr(feature="serde", derive(Deserialize, Serialize))]
     #[derive(Clone, Debug, Eq, PartialEq)]
     pub struct Transaction {
         /// Chain tag
+        #[serde(rename="chainTag")]
         pub chain_tag: u8,
         /// Previous block reference
         ///
         /// First 4 bytes (BE) are block height, the rest is part of referred block ID.
+        #[serde(rename="blockRef")]
         pub block_ref: u64,
         /// Expiration (in blocks)
         pub expiration: u32,
         /// Vector of clauses
         pub clauses: Vec<Clause>,
         /// Coefficient to calculate the gas price.
+        #[serde(rename="gasPriceCoef")]
         pub gas_price_coef: u8,
         /// Maximal amount of gas to spend for transaction.
         pub gas: u64,
         /// Hash of transaction on which current transaction depends.
         ///
         /// May be left unspecified if this functionality is not necessary.
+        #[serde(rename="dependsOn")]
         pub depends_on: Option<U256> => AsBytes<U256>,
-        /// Transaction nonce
+        /// Transaction nonceserde_as
         pub nonce: u64,
         /// Reserved fields.
         pub reserved: Option<Reserved> => AsVec<Reserved>,
@@ -42,6 +50,7 @@ rlp_encodable! {
         ///
         /// For VIP-191 transactions, this would be a simple concatenation
         /// of two signatures.
+        #[cfg_attr(feature="serde", serde(with = "serde_with::As::<Option<crate::utils::unhex::Hex>>"))]
         pub signature: Option<Bytes> => Maybe<Bytes>,
     }
 }
@@ -248,13 +257,16 @@ impl Transaction {
 
 rlp_encodable! {
     /// Represents a single transaction clause (recipient, value and data).
-    #[derive(Clone, Debug, Eq, PartialEq)]
+    #[serde_with::serde_as]
+    #[cfg_attr(feature="serde", derive(Deserialize, Serialize))]
+    #[cfg_attr(feature="serde", derive(Clone, Debug, Eq, PartialEq))]
     pub struct Clause {
         /// Recipient
         pub to: Option<Address> => AsBytes<Address>,
         /// Amount of funds to spend.
         pub value: U256,
         /// Contract code or other data.
+        #[cfg_attr(feature="serde", serde(with = "serde_with::As::<crate::utils::unhex::Hex>"))]
         pub data: Bytes,
     }
 }
@@ -295,11 +307,17 @@ impl Clause {
 }
 
 /// Represents a transaction's ``reserved`` field.
+#[cfg_attr(feature = "serde", serde_with::serde_as)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct Reserved {
     /// Features to enable (bitmask).
     pub features: u32,
     /// Currently unused field.
+    #[cfg_attr(
+        feature = "serde",
+        serde(with = "serde_with::As::<Vec<crate::utils::unhex::Hex>>")
+    )]
     pub unused: Vec<Bytes>,
 }